fix: :bug: Again, type issue, and some UEFI code

+++ library/src/loader_main/mod.rs
@@ -110,8 +110,8 @@ fn get_framebuffer() -> Framebuffer {
     // Init the VBE info
     let vbe = VBEInfo::load(0x10000); // Put in fixed address in stage2
     let fb_addr: u64 = 0xffff800040000000; // Mapped
-    let width: u32 = vbe.x_resolution.into();
-    let height: u32 = vbe.y_resolution.into();
+    let width: u64 = vbe.x_resolution.into();
+    let height: u64 = vbe.y_resolution.into();
     let bpp: u64 = (vbe.bits_per_pixel / 8).into();
     let pitch: u64 = vbe.bytes_per_scan_line.into();
